Group iOS progress files
