+++ apps/ios/Flashcards/Flashcards/Progress/Model/ProgressCoreTypes.swift
@@ -1,9 +1,6 @@
 import Foundation
-import SwiftUI
 
 let progressDaysPerWeek: Int = 7
-private let progressStreakWeekCount: Int = 5
-let reviewProgressBadgeOverflowThreshold: Int = 99
 
 struct ProgressDay: Codable, Hashable, Identifiable, Sendable {
     let date: String
@@ -275,170 +272,6 @@ enum ProgressSourceState: String, Codable, Hashable, Sendable {
     case serverBaseWithPendingLocalOverlay = "server_base_with_pending_local_overlay"
 }
 
-struct ProgressScopeKey: Codable, Hashable, Sendable {
-    let cloudState: CloudAccountState?
-    let linkedUserId: String?
-    /// Tracks the canonical cached workspace membership that contributes to aggregated progress.
-    /// Switching the active workspace keeps this stable, but create/delete/merge membership changes rotate the scope.
-    let workspaceMembershipKey: String
-    let timeZone: String
-    let from: String
-    let to: String
-
-    var storageKey: String {
-        let cloudStateKey = self.cloudState?.rawValue ?? "none"
-        let linkedUserIdKey = self.linkedUserId ?? "none"
-        return [
-            cloudStateKey,
-            linkedUserIdKey,
-            self.workspaceMembershipKey,
-            self.timeZone,
-            self.from,
-            self.to,
-        ].joined(separator: "|")
-    }
-}
-
-struct ProgressSummaryScopeKey: Codable, Hashable, Sendable {
-    let cloudState: CloudAccountState?
-    let linkedUserId: String?
-    let workspaceMembershipKey: String
-    let timeZone: String
-    /// Summary fields such as hasReviewedToday and currentStreakDays are relative to a local "today".
-    /// Keep the cache keyed by that local date so yesterday's summary is never reused after midnight.
-    let referenceLocalDate: String
-
-    var storageKey: String {
-        let cloudStateKey = self.cloudState?.rawValue ?? "none"
-        let linkedUserIdKey = self.linkedUserId ?? "none"
-        return [
-            cloudStateKey,
-            linkedUserIdKey,
-            self.workspaceMembershipKey,
-            self.timeZone,
-            self.referenceLocalDate,
-        ].joined(separator: "|")
-    }
-}
-
-struct ReviewScheduleScopeKey: Codable, Hashable, Sendable {
-    let cloudState: CloudAccountState?
-    let linkedUserId: String?
-    let workspaceMembershipKey: String
-    let timeZone: String
-    /// Schedule buckets are relative to the user's local "today".
-    /// Keep the cache keyed by that local date so "Today" rotates cleanly after midnight.
-    let referenceLocalDate: String
-
-    var storageKey: String {
-        let cloudStateKey = self.cloudState?.rawValue ?? "none"
-        let linkedUserIdKey = self.linkedUserId ?? "none"
-        return [
-            cloudStateKey,
-            linkedUserIdKey,
-            self.workspaceMembershipKey,
-            self.timeZone,
-            self.referenceLocalDate,
-        ].joined(separator: "|")
-    }
-}
-
-struct ProgressCalendarDay: Hashable, Identifiable, Sendable {
-    let date: Date
-    let localDate: String
-    let reviewCount: Int
-    let isToday: Bool
-    let isFuturePlaceholder: Bool
-    let dayNumber: Int
-
-    var id: String {
-        self.localDate
-    }
-}
-
-struct ProgressCalendarWeek: Hashable, Identifiable, Sendable {
-    let days: [ProgressCalendarDay]
-
-    var id: String {
-        guard let firstDay = self.days.first else {
-            preconditionFailure("Progress calendar week must contain at least one day")
-        }
-
-        return firstDay.localDate
-    }
-}
-
-struct ProgressChartDay: Hashable, Identifiable, Sendable {
-    let date: Date
-    let localDate: String
-    let reviewCount: Int
-    let isToday: Bool
-
-    var id: String {
-        self.localDate
-    }
-}
-
-struct ProgressChartData: Hashable, Sendable {
-    let chartDays: [ProgressChartDay]
-}
-
-struct ProgressSnapshot: Hashable, Sendable {
-    let scopeKey: ProgressScopeKey
-    let summary: ProgressSummary
-    let chartData: ProgressChartData
-    let summarySourceState: ProgressSourceState
-    let seriesSourceState: ProgressSourceState
-    let isApproximate: Bool
-    let generatedAt: String?
-}
-
-struct ReviewScheduleSnapshot: Hashable, Sendable {
-    let scopeKey: ReviewScheduleScopeKey
-    let schedule: UserReviewSchedule
-    let sourceState: ProgressSourceState
-    let isApproximate: Bool
-    let generatedAt: String?
-}
-
-struct ReviewProgressBadgeState: Hashable, Sendable {
-    let streakDays: Int
-    let hasReviewedToday: Bool
-    let isInteractive: Bool
-}
-
-struct ReviewProgressBadgePresentation {
-    let iconSystemName: String
-    let borderColor: Color
-    let iconColor: Color
-    let textColor: Color
-}
-
-func makeEmptyReviewProgressBadgeState() -> ReviewProgressBadgeState {
-    ReviewProgressBadgeState(
-        streakDays: 0,
-        hasReviewedToday: false,
-        isInteractive: true
-    )
-}
-
-func makeReviewProgressBadgePresentation(badgeState: ReviewProgressBadgeState) -> ReviewProgressBadgePresentation {
-    ReviewProgressBadgePresentation(
-        iconSystemName: badgeState.hasReviewedToday ? "flame.fill" : "flame",
-        borderColor: badgeState.hasReviewedToday ? .accentColor.opacity(0.55) : .gray.opacity(0.35),
-        iconColor: badgeState.hasReviewedToday ? .accentColor : .gray,
-        textColor: badgeState.hasReviewedToday ? .primary : .secondary
-    )
-}
-
-func formatReviewProgressBadgeValue(badgeState: ReviewProgressBadgeState) -> String {
-    if badgeState.streakDays > reviewProgressBadgeOverflowThreshold {
-        return "\(reviewProgressBadgeOverflowThreshold)+"
-    }
-
-    return badgeState.streakDays.formatted()
-}
-
 enum ProgressPresentationError: LocalizedError {
     case duplicateDay(String)
     case invalidLocalDate(String)
@@ -481,143 +314,6 @@ enum ProgressPresentationError: LocalizedError {
             return "Review schedule metadata mismatched the current scope. Expected \(expectedTimeZone), received \(actualTimeZone)."
         }
     }
-}
-
-func makeReviewProgressBadgeState(progressSnapshot: ProgressSnapshot?) -> ReviewProgressBadgeState {
-    guard let progressSnapshot else {
-        return makeEmptyReviewProgressBadgeState()
-    }
-
-    return ReviewProgressBadgeState(
-        streakDays: progressSnapshot.summary.currentStreakDays,
-        hasReviewedToday: progressSnapshot.summary.hasReviewedToday,
-        isInteractive: true
-    )
-}
-
-func makeReviewProgressBadgeState(summary: ProgressSummary) -> ReviewProgressBadgeState {
-    ReviewProgressBadgeState(
-        streakDays: summary.currentStreakDays,
-        hasReviewedToday: summary.hasReviewedToday,
-        isInteractive: true
-    )
-}
-
-func makeProgressSnapshot(
-    summary: ProgressSummary,
-    series: UserProgressSeries,
-    scopeKey: ProgressScopeKey,
-    summarySourceState: ProgressSourceState,
-    seriesSourceState: ProgressSourceState,
-    calendar: Calendar
-) throws -> ProgressSnapshot {
-    let timeline = try makeValidatedProgressTimeline(
-        series: series,
-        scopeKey: scopeKey,
-        calendar: calendar
-    )
-    let todayLocalDate = series.to
-
-    let chartDays = timeline.map { timelineDay in
-        ProgressChartDay(
-            date: timelineDay.date,
-            localDate: timelineDay.localDate,
-            reviewCount: timelineDay.reviewCount,
-            isToday: timelineDay.localDate == todayLocalDate
-        )
-    }
-    let chartData = ProgressChartData(
-        chartDays: chartDays
-    )
-
-    return ProgressSnapshot(
-        scopeKey: scopeKey,
-        summary: summary,
-        chartData: chartData,
-        summarySourceState: summarySourceState,
-        seriesSourceState: seriesSourceState,
-        isApproximate: summarySourceState == .localOnly || seriesSourceState == .localOnly,
-        generatedAt: series.generatedAt
-    )
-}
-
-func makeReviewScheduleSnapshot(
-    schedule: UserReviewSchedule,
-    scopeKey: ReviewScheduleScopeKey,
-    sourceState: ProgressSourceState
-) throws -> ReviewScheduleSnapshot {
-    try validateReviewSchedule(schedule: schedule, scopeKey: scopeKey)
-
-    return ReviewScheduleSnapshot(
-        scopeKey: scopeKey,
-        schedule: schedule,
-        sourceState: sourceState,
-        isApproximate: sourceState != .serverBase,
-        generatedAt: schedule.generatedAt
-    )
-}
-
-private struct ProgressTimelineDay: Hashable, Sendable {
-    let date: Date
-    let localDate: String
-    let reviewCount: Int
-}
-
-private func makeProgressTimeline(
-    series: UserProgressSeries,
-    calendar: Calendar
-) throws -> [ProgressTimelineDay] {
-    let startDate = try progressDate(localDate: series.from, calendar: calendar)
-    let endDate = try progressDate(localDate: series.to, calendar: calendar)
-
-    guard startDate <= endDate else {
-        throw ProgressPresentationError.invalidRange(series.from, series.to)
-    }
-
-    var reviewCountsByLocalDate: [String: Int] = [:]
-    for day in series.dailyReviews {
-        _ = try progressDate(localDate: day.date, calendar: calendar)
-
-        guard day.reviewCount >= 0 else {
-            throw ProgressPresentationError.negativeReviewCount(day.date, day.reviewCount)
-        }
-
-        if reviewCountsByLocalDate.updateValue(day.reviewCount, forKey: day.date) != nil {
-            throw ProgressPresentationError.duplicateDay(day.date)
-        }
-    }
-
-    var timeline: [ProgressTimelineDay] = []
-    var currentDate = startDate
-    while currentDate <= endDate {
-        let localDate = progressLocalDateString(date: currentDate, calendar: calendar)
-        timeline.append(
-            ProgressTimelineDay(
-                date: currentDate,
-                localDate: localDate,
-                reviewCount: reviewCountsByLocalDate[localDate] ?? 0
-            )
-        )
-
-        guard let nextDate = calendar.date(byAdding: .day, value: 1, to: currentDate) else {
-            throw ProgressPresentationError.invalidRange(series.from, series.to)
-        }
-        currentDate = nextDate
-    }
-
-    return timeline
-}
-
-func validateProgressSeries(
-    series: UserProgressSeries,
-    scopeKey: ProgressScopeKey,
-    calendar: Calendar
-) throws {
-    _ = try makeValidatedProgressTimeline(
-        series: series,
-        scopeKey: scopeKey,
-        calendar: calendar
-    )
 }
 
 func makeProgressSummary(
@@ -737,49 +433,6 @@ func validateReviewSchedule(
     }
 }
 
-func makeProgressPresentationCalendar(
-    timeZoneIdentifier: String,
-    userCalendar: Calendar
-) throws -> Calendar {
-    guard let timeZone = TimeZone(identifier: timeZoneIdentifier) else {
-        throw ProgressPresentationError.invalidTimeZone(timeZoneIdentifier)
-    }
-
-    var calendar = Calendar(identifier: .gregorian)
-    calendar.locale = Locale.autoupdatingCurrent
-    calendar.timeZone = timeZone
-    calendar.firstWeekday = userCalendar.firstWeekday
-    calendar.minimumDaysInFirstWeek = userCalendar.minimumDaysInFirstWeek
-    return calendar
-}
-
-func validateProgressSeriesMetadata(
-    series: UserProgressSeries,
-    scopeKey: ProgressScopeKey
-) throws {
-    guard
-        series.timeZone == scopeKey.timeZone,
-        series.from == scopeKey.from,
-        series.to == scopeKey.to
-    else {
-        throw ProgressPresentationError.seriesMetadataMismatch(
-            expected: scopeKey,
-            actualTimeZone: series.timeZone,
-            actualFrom: series.from,
-            actualTo: series.to
-        )
-    }
-}
-
-private func makeValidatedProgressTimeline(
-    series: UserProgressSeries,
-    scopeKey: ProgressScopeKey,
-    calendar: Calendar
-) throws -> [ProgressTimelineDay] {
-    try validateProgressSeriesMetadata(series: series, scopeKey: scopeKey)
-    return try makeProgressTimeline(series: series, calendar: calendar)
-}
-
 private func calculateProgressCurrentStreakDays(
     reviewDates: Set<String>,
     todayLocalDate: String
@@ -837,70 +490,7 @@ private func progressTimeZoneLocalDateString(
     return formatter.string(from: date)
 }
 
-func makeProgressStreakWeeks(
-    chartDays: [ProgressChartDay],
-    rangeStartLocalDate: String,
-    todayLocalDate: String,
-    calendar: Calendar
-) throws -> [ProgressCalendarWeek] {
-    let today = try progressDate(localDate: todayLocalDate, calendar: calendar)
-
-    guard let currentWeekInterval = calendar.dateInterval(of: .weekOfYear, for: today) else {
-        throw ProgressPresentationError.invalidRange(rangeStartLocalDate, todayLocalDate)
-    }
-
-    let currentWeekStart = calendar.startOfDay(for: currentWeekInterval.start)
-    let streakDayCount = progressDaysPerWeek * progressStreakWeekCount
-
-    guard let streakStart = calendar.date(byAdding: .day, value: -(streakDayCount - progressDaysPerWeek), to: currentWeekStart) else {
-        throw ProgressPresentationError.invalidRange(rangeStartLocalDate, todayLocalDate)
-    }
-
-    let chartDaysByLocalDate = Dictionary(uniqueKeysWithValues: chartDays.map { ($0.localDate, $0) })
-    let streakDays = try (0 ..< streakDayCount).map { offset in
-        guard let rawDate = calendar.date(byAdding: .day, value: offset, to: streakStart) else {
-            throw ProgressPresentationError.invalidRange(rangeStartLocalDate, todayLocalDate)
-        }
-
-        let date = calendar.startOfDay(for: rawDate)
-        let localDate = progressLocalDateString(date: date, calendar: calendar)
-        let isFuturePlaceholder = date > today
-        let reviewCount: Int
-
-        if isFuturePlaceholder == false {
-            guard let chartDay = chartDaysByLocalDate[localDate] else {
-                throw ProgressPresentationError.invalidRange(rangeStartLocalDate, todayLocalDate)
-            }
-
-            reviewCount = chartDay.reviewCount
-        } else {
-            reviewCount = 0
-        }
-
-        return ProgressCalendarDay(
-            date: date,
-            localDate: localDate,
-            reviewCount: reviewCount,
-            isToday: localDate == todayLocalDate,
-            isFuturePlaceholder: isFuturePlaceholder,
-            dayNumber: calendar.component(.day, from: date)
-        )
-    }
-
-    return stride(from: 0, to: streakDays.count, by: progressDaysPerWeek).map { startIndex in
-        ProgressCalendarWeek(days: Array(streakDays[startIndex ..< startIndex + progressDaysPerWeek]))
-    }
-}
-
-func progressChartUpperBound(maximumReviewCount: Int) -> Int {
-    guard maximumReviewCount > 0 else {
-        return 1
-    }
-
-    return max(1, Int(ceil(Double(maximumReviewCount) * 1.1)))
-}
-
-private func progressDate(localDate: String, calendar: Calendar) throws -> Date {
+func progressDate(localDate: String, calendar: Calendar) throws -> Date {
     guard let date = progressStrictDate(localDate: localDate, calendar: calendar) else {
         throw ProgressPresentationError.invalidLocalDate(localDate)
     }
