+++ apps/ios/Flashcards/Flashcards/Progress/Presentation/ProgressFormatting.swift
